drivers: power: ltc4162l: fix incompatible-pointer-types for GCC 14+
GCC 14 makes -Wincompatible-pointer-types an error. Size the uval read
locals as uint16_t to match the accessors they are passed to, and widen the
store value to int32_t to match the channel store helper's expected pointer
width.

+++ drivers/power/ltc4162l/iio_ltc4162l.c
@@ -131,7 +131,7 @@ static int ltc4162l_ch_attr_show(void *ddev, char *buf,
 	struct ltc4162l_iio_device *dev = ddev;
 	int ret;
 	int32_t val;
-	uint32_t uval;
+	uint16_t uval;
 
 	switch (priv) {
 	case LTC4162L_BATTERY_COUNT:
@@ -216,7 +216,7 @@ static int ltc4162l_ch_attr_store(void *ddev, char *buf,
 				  intptr_t priv)
 {
 	struct ltc4162l_iio_device *dev = ddev;
-	int16_t val;
+	int32_t val;
 	int ret;
 
 	switch (priv) {
@@ -296,7 +296,7 @@ static int ltc4162l_read_raw(void *ddev, char *buf, uint32_t len,
 	struct ltc4162l_iio_device *dev = ddev;
 	int ret;
 	int32_t val;
-	uint32_t uval;
+	uint16_t uval;
 
 	switch (priv) {
 	case LTC4162L_VBAT_VAL:
